2018 Day 6 Part 1
Man, this solution is so whack. Honestly, I probably could have just ignored any set which touches the perimeter of the bounding box of the points and treated them as infinite. However, I wasn't going to sit there and prove that that holds, so I instead went with the hacky solution of assuming that with a 9x larger grid (tripling each side length) any finite set will have the same size and inifinite sets will be larger. (This was easier to code too.) It's definitely *not* fast, but it would have taken way longer to do "properly", and this did give me the answer.

Would have been 27th on the leaderboard

Time: 0:10:58.846114