2018 Day 7 Part 1
I had a few dumb false starts here. This isn't a strict topo sort, nor is it working through the dependencies of the lowest step in ascending order to start step A fastest. At each point we need to find the lowest step that we can start *now* and do it then! Not a complex idea, but it took me at least 3 tries to actually work on the right problem.

Once I was on the right track it fell in place quickly, at least.

Would have been 77th or 78th on the leaderboard.

Time: 0:08:17.975017