2018 Day 8 Part 2
This was a fun twist, metadata is now a set of indices for the child evaluations! Not too complex to do...though I did waste time due to forgetting to increment my counts for each time a child node was referenced. That was a dumb waste.

Would have been 58th, 59th, or 60th on the leaderboard.

Time: 0:09:50.840189