2018 Day 9 Part 1
This problem screams for a doubly linked list. However, implementing a linked list takes time and is icky. The collections.deque class can also be pretty useful in these sorts of cases, if you know how to use it.

The problem is that I am not too familiar with its API so I spent the first few minutes just playing with deque in the REPL to understand what I needed from it. Once that was done it was generally pretty easy to code up the logic.

I did have an off by one error in where I was inserting the marbles, but that was easy to correct once I figured out what was going on.

Would have been 26th on the leaderboard. Not bad for me really not being that familiar with collections.deque!

Time: 0:10:44.560985