2018 Day 9 Part 2
And part 2 is just pushing it to the extreme. I do notice a slight bit of lag now, but nothing too major. It still finishes in about a second which isn't bad at all.

Would have been 7th on the leaderboard! Wow, that's *really* not bad for me not really having used deque before!

Time: 0:11:42.652605