2018 Rewrite Day 11
It turns out that my initial idea was slower, though in a way on the right track. What I needed was 2D partial sums where the partial sums all start at the origin and include a rectangle. It's possible to compute a block's power with 4 2D partial sums to eliminate sections above and to the left.

Doing this doesn't speed things up that much (maybe 30-40% down to ~7 seconds), but it's a *whole* lot simpler and feels a lot better.