2018 Day 12 Part 2
Hoo boy. 50 billion iterations pretty clearly screams "I end up in a cycle, find it". The trick was in finding the pattern and figuring out how to code it!

In the end the pattern ended up falling into a cycle that shifts after a while. (I found that it is 100% stable even, just moving to the right each generation. That's probably true for all inputs, but I'm not going to hardcode that.)

The tricky thing here is that with the pattern shifting, the representation needs to change to be the plants and their starting offset, otherwise we can't detect a matching pattern. This took quite a bit to get working, and honestly I don't even remember how much of the slowdown was bugs vs the complexity of managing data. I do know that i hit some bugs though, and I'm super glad that I got the answer for 20k generations the slow way because that let me test my skip-ahead code until I got it right without pinging the server!

Would have missed the leaderboard by 6 minutes and 8 seconds, but that's fine.

Time: 0:33:50.250199