2018 Refactor Day 12
While I initially wrote this using lists, sets, and tuples as if it were a Game of Life problem, it really makes more sense to use strings. That way string slicing gives the key to the transition table directly. Not only that, but the str.index method directly gives the necessary adjustment for a resultant pot string which is very convenient.

Furthermore, If I actually use the sequence array then I can complete the jump ahead immediately instead of having a secondary stepping loop to complete the generations. It's a bit minor (especially with the stepping logic being so clean now) but it's still nicer.