2018 Day 13 Part 2
Okay, that was nasty. This problem 100% should have had better examples to exemplify that carts can crash into other carts that have not yet moved. In fact I'd go so far as to say that the problem is intentionally written to mislead people. Even worse, for some people their input runs forever if you miss that edge case and for others (like me) the problem converges on a "wrong" answer if you miss that rule.

I went over my code countless times to find no bug, and even rewrote swaths of it just in case. In the end it was just the problem being bad. (Because seriously: This grid is *way* too big to be debugging, there is *no* way we should be expected to have to debug it to discover unstated rules.)

Almost twice the 100th time, and I had to see the />>-\ edge case on the Reddit to realize what was missing (from the problem description).

Time: 1:27:48.782832