2018 Day 14 Part 2
Oo, that's nice. I was flirting with creating a state machine similar to a regex to track when the "win" condition is reached by just passing in the new recipes one at a time, but that sounds complicated. Keeping a short list of the most recent recipe scores is much easier to code. It doesn't take too long either, which works for me. (In fact, generating the recipes takes way longer, I'd initially thought that I was being slowed down by my approach!)

Would have been 20th on the leaderboard! I'm surprised I jumped up so far.

Time: 0:10:51.320884