2018 Day 15 Part 1
Wow. That was a big problem, but honestly, I rather enjoyed it. Coding up a full battle simulation with very explicit rules was rather fun. I did initially get the attack priority wrong (missing that lower hp targets were prioritized), but that was really the only bug I had.

What was interesting was coming up with the systems to manage the game state on the fly while trying to solve this fast. But I *really* enjoyed that!

Would have been 9th on the leaderboard?! Wow! I thought I was doing this slowly!

Time: 0:49:43.187979