2018 Day 16 Part 1
Now this was interesting. The longest bit here was coding all the instructions, and a lot of that was due to needing to validate that the arguments were in range and handling when they're invalid. (Technically possible for some of the immediate instructions, though that didn't seem to happen in practice.)

Once that was done and stored in a dictionary though it's super easy to just loop through all instruction candidates, try them, and see if the sample output matches the instruction output. Pretty interesting, and I'm sure I've set up well for part 2 here.

Would have been 31st on the leaderboard.

Time: 0:15:41.851030