2018 Day 17 Part 1
Man, 2018 must have been the year of the sim. Now we're simulating crude water physics. My approach here is somewhat crude, but it works. The biggest flaw is that I recurse a *ton* here (beyond the default Python stack limit) which I hacked around, but that can be cleaned up another time.

Thankfully our ultimate question for every cell touched by water is simple: Is it still or flowing. If it's supported by clay or still water and is surrounded by clay it's still, otherwise it's flowing. Whenever we're unsure we just have to check below and see if that cell is flowing or still.

I did have a couple of bugs here, but this was a very visual problem so all I needed to do was format the information, dump it to a file, and load it in notepad. Both bugs (both bugs in the lateral flow of water) were pretty quickly apparent!

And wow! This would have been first on the leaderboard! In fact, this time beats the top submission time by 1 minute and 54 seconds!

Time: 0:30:49.771919