2018 Day 18 Part 2
Ah, theres' the twist. 1 BILLION iterations of this. My stepping algorithm wasn't too fast, but thankfully with such a high number of iterations to go through it was pretty obvious that there were going to be repeated states. The tricky thing was detectin them! I ended up deciding to expose an as_str method on my grid class since that's basically like printing and it gives me a unique key for each state. Then it's just standard state recognition and jump-ahead when a duplicate state is seen!

Would have been 41st on the leaderboard. I'll take it, honestly.

Time: 0:16:17.182380