2018 Day 19 Part 1
We're bringing back the day 16 instruction set. Fun. It took a bit to get the instruction pointer bit working, but that was due to me missing the rules about when the instruction pointer was written to the register or read from the register. Beyond that, pretty simple.

Would have barely made the leaderboard at either 98th, 99th, or 100th. Oof.

Time: 0:12:11.299619