2018 Day 20 Part 1
Oh now this is an interesting problem. I at first thought that the regex library was going to be helpful, but that's almost certainly not a good idea here. The matching patterns would be looooong.

Anyway, parsing the regex into a tree structure wasn't too bad. Then when that was done it's just a matter of walking the regex and keeing track of the superposition of possible locations. With that it's pretty easy to construct the map. (Well, the list of doors in each room, technically.)

With that it was pretty trivial to convert to a graph for my lib.graph.longest_path_length function to give me the answer!

Would have been 46th on the leaderboard.

Time: 0:33:37.329909