2018 Day 20 Part 2
Well that's easy. I have the graph already so it's just a matter of calling lib.graph.all_reachable and filtering based on the distance.

The *only* hiccup is that the graph format is slightly different compared to longest_path_length. I really need to make lib.graph consistent...

Would have been 40th on the leaderboard.

Time: 0:35:00.912731