2018 Day 21 Part 1
Yuck. Stop with the decompilation problems. These aren't satisfying.

Granted, I didn't truly decompile here, I just traced through the simulation and looked for which instructions could cause a halt. After that it was just a matter of seeing what the registers were for the condition near the halt and picking the first one.

Would have been 40th on the leaderboard.

Time: 0:14:44.005522