2018 Day 21 Part 2
And now to fully decompile. See how there's no changed code? This is a bad problem. No information is actually retained in the source code from this worthless decompilation problem.

Anyway. This seems to be some RNG which tests a bunch of numbers in sequence. If regs[0] matches any of them then the program halts, otherwise it doesn't.

I'm going to see if I can solve this programmatically, that would be so much more satisfying...

Would have been 41st on the leaderboard.

Time: 0:43:24.728373