2018 Rewrite Day 21
Using a special type for register 0 I can detect any comparisons with register 0 and enter a "quantum" superstate. Then whenever the ip register is set to a "quantum" superstate value I can check to see if one would halt. If so, whatever value matches the equality test would have halted.

Then if we were to return the same value a second time we can just quit since we've entered a big cycle. (Technically I should check the full register state but that's unnecessarily complicated.)

The only additional thing here is optimizing the inner floor division loop. Once that's done both parts 1 and 2 finish deterministically and quickly.

So much nicer than a decompiled program in notepad which is discarded. *So* much nicer.