2018 Day 22 Part 1
Okay, that was a bit more tricksy than I expected. I completely missed that the geo index was recursing on the *erosion* index which was my biggest time loss. But before that I wasted time not realizing how gigantic the numbers were going to be. (I was a little wary of the multiplication, but for some reason didn't immediately do modular arithmetic...) Once both of those were fixed though it solved just fine.

Sidenote: I briefly thought that we needed to do the 2D partial sums again, not thinking about the fact that there are no negative risks so that was just dumb.

Would have missed the leaderboard by just over 8 minutes. Ouch.

Time: 0:22:55.415845