2018 Day 22 Part 2
And now for a graph search. I expected as much, honestly. My lib.graph library is quite well prepared for this so it was just a matter of getting the neighbor function correct.

I did initially mistake what the end state was (thinking we just needed to reach the target with any tool) so I got the wrong answer initially. (And wasted time with dijkstra_length_fuzzy_end's end function!)

I *was* a little worried when my graph search took more than a couple seconds to find the answer (taking about 12 seconds) but it just was a large state space to search through.

Would have jumped up to 33rd on the leaderboard! Wow. This would have been a really solid placement had I not botched part 1.

Time: 0:37:08.617164