2018 Day 23 Part 1
Wow did I botch this badly. First I took the bot with the least signal strength, then I neglected to include the bot in its own radius. (I'd argue that a bot shouldn't be "in range" of itself, but the problem was clear on that.)

Would still have been 50th on the leaderboard?! Wow. That's...kind of sad actually. If I *hadn't* botched this part so badly and had two minute-long waits I'd have been 2nd.

Time: 0:05:14.323921