2018 Day 23 Part 2
Now this is an interesting problem. I'm honestly not sure if I did this right at all, but it worked.

My first inclination was to use the average of the bots positions as the starting point, but then I realized that bots with larger ranges have way more flexibility. As such, a weighted average gives a better starting point, and for the example is nearly the answer. However, searching around the current guess is complicated and if done poorly can be very slow since there's a lot of ground to cover.

I think that the approach I came up with is related to gradient ascent? I'm just trying all directions around the current point at different factors and seeing if I can make progress. Honestly it's pretty jank and haphazard, but it does give me the right answer.

Would have been an insane 8th place on the leaderboard! Wow!

Time: 0:30:07.082967