2019 Day 2 Part 1
Oops! I didn't mean to start an Intcode problem tonight! Thankfully the first one is very basic, just 3 opcodes to deal with.

I'll definitely need to be refactoring this for later in 2019, but that can happen later.

Would have been 26th, but I did the problem in 2019 so this doesn't exactly count.

Time: 0:04:29.468329