2019 Day 2 Part 2
Ah right, the noun/verb search. This took a bit longer than I wanted in the first pass, but it turns out that that was due to using the parse library 10k times instead of faster program parsing methods.

All in all pretty basic. I'll maybe try to refactor the Intcode stuff for the future tomorrow.

Would have been 9th, but I did the problem in 2019 so this doesn't exactly count.

Time: 0:06:25.441665