2019 Day 3 Part 2
Ah, and we're counting how far along the wires we are at the intersection. Not too bad, just a bit of extra logic in the parse_wire function to count steps and a different min comprehension.

Would have been 7th on the leaderboard (or more like 18th with the minute timeout), but I did the problem in 2019 so this doesn't exactly count.

Time: 0:08:05.382505