2019 Refactor Intcode VM
Having both synchronous and asynchronous running modes is silly when most of the time I'll have inputs and outputs to deal with. Instead it makes more sense for run to start the VM in a backgrond thread and set up channels. This marginally complicates day 2's solution but will be simpler in the long run. However, adding a wait_for_close method on the channels simplifies day 2 nearly to how the original solution looked.

Additionally, calling take_num and take_mode_num is janky. Having a params function in the Intcode VM makes param retrieval much simpler for each operation. It also reduces the risk of me making coding errors down the line (since I know there's a third parameter mode that will be introduced and I remember bugs in my go Intcode VM in 2019).