2019 Day 6 Part 2
YOU and SAN will eventually have *some* common indirect orbit in the chain, even if it's COM itself. The number of orbit transfers you need to make to reach SAN is the same as if you and SAN both transferrd to your most common orbit, so it's a simple matter of finding that common orbit and counting!

Would have been 7th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count.

Time: 0:06:21.765048