2019 Day 7 Part 1
Oh hey, we're chaining instances of Intcode programs and giving them different inputs. Fun, and also not too bad to do. Even without channels shenanigans this is easy to do since it's running the VM instances in sequence.

Would have been 5th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count. (Plus I specifically designed the Intcode VM fo chaining like this which was a surprise back when this problem landed!)

Time: 0:04:19.536595