2019 Day 7 Part 2
Ah, right. And now the last output can feed back on the first input. As a result the program needs to run in the background. Good thing I already have that happening, though in 2019 this was a complete surprise.

One thing I don't like is the manual linking of input and output channels. In fact, it may make sense to have the intcode VM allow clients to pass in the input channel. That would simplify the chaining here a good bit.

Would have been first on the leaderboard, but I did the problem in 2019 so this doesn't exactly count. (Plus I specifically designed the Intcode VM fo chaining like this which was a surprise back when this problem landed!)

Time: 0:10:09.461879