2019 Refactor Intcode VM
Allow passing in an input channel to run. This way chaining programs is nice and easy when desired. (Granted if required initialization inputs need to be sent then there could be synchronization issues, but in this case the previous amplifier won't output until we give it an input signal so there's no chance of a race condition.)