2019 Refactor Day 8
Thinking about it, it makes more sense to have a partition helper to take a list of numbers and partition it. The image composition can happen all as a list, then the composed list can be partitioned into the 2D image.

Plus the white cell list can be generated with a list comprehension.