2019 Day 9 Part 1
Hey, here's the final parameter mode. It took a bit of doing to handle the output relative parameter mode correctly, but even that wasn't too bad.

Also, here's the introduction of infinite memory. This memory growing scheme may be inefficient, I'm not sure about the efficiency of Python list growth. But it works and is enclosed in the params function so at least it just works for everything.

I remember this hit me hard in 2019, I'm glad things just worked this time. (If I remember correctly it may have even been a bug in a previous instruction!)

Would have been 30th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count.

Time: 0:08:16.146209