2019 Day 10 Part 2
Ah, and now we need to know more about the angles. Namely, which asteroids are visible at any given angle and what order the angles come in.

I'm sure there's a much smarter way to sort the angles, but breaking the angles into right half and left half lets me sort based on dy/dx to get clockwise order. Then it's just a matter of stitching those together with the two vertical angles and I've got the full sweep. When cycling through the angles, we can just pick the asteroid with the smallest manhattan distance to the lookout asteroid. Actually pretty easy!

Would have dropped down to 8th, but I did the problem in 2019 so this doesn't exactly count. (Still though, that's high!)

Time: 0:23:36.818431