2019 Rewrite Day 10 Part 2
It turns out that I had a bug in part 2! I was (stupidly) thinking that a dx,dy of 0,1 was pointing up, forgetting that low y is up. That means that 0,-1 is up and I had the intended sorting all mixed up for my angle order.

Additionally, I forgot that dx would be negative for left side angles so left_side_angles was sorted in the opposite order that I intended, correcting for the above mistake.

Thankfully the right asteroid was in the left quadrant so these bugs canceled out. However, that's really icky behavior so I reworked the angle sorting to be based on math.atan2 which is smarter than me. The only issue is that it doesn't put the up angle first in the list so the angle order list needs to be shifted.

One other small improvement here is sorting the asteroids based on their manhattan distance from the station in determine_asteroid_visibility. This makes the destruction order much easier to compute since we just pop the 0th item from the list until it's empty.