2019 Day 11 Part 1
Hey, painting tiles on a grid is fun. Pretty easy to whip up too, especially with the channel library I set up and am using for my Intcode VM.

Would have been 5th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count. Plus I made a good channel library. That aside, that's still a high rank!

Time: 0:05:07.803752