2019 Day 12 Part 2
And part 2. I can't forget the trick to this even if I tried. Each component of the moon's properties acts independently so we can simulate each set of components individually to get the x, y, and z cycle congruencies. It still takes a moment, but it finishes.

Then once we have those congruencies we just need to find out how long until those congruencies line up. I was going to use lib.math.chinese_remainder for this (to not have to hardcode any assumptions about the offsets in the congruencies) but that returns 0 because the offsets are all 0 in practice! As such I had to whip up the computation manually with math.gcd. I'll expand lib.math.chinese_congruencies in a follow-up.

Would have been 21st, but I did the problem in 2019 so this doesn't exactly count. Especially since I already know the trick to this problem!

Time: 0:22:55.327180