2019 Refactor Day 12 Part 2
With a quick adjustment to lib.math.chinese_remainder it can now give answers that respect a minimum (in this case 1). That way when given the congruences it'll give the minimum non-zero match, exactly what we want here. The problem says the first state that matches a previous state, not necessarily that matches the start state, after all.

Then there's just some minor cleanup in the solution itself. Nothing to write home about.