2019 Day 14 Part 1
Oh hey, we're creating fuel from ore. I remember getting tripped up on this big time due to forgetting about surplus ingredients so I made sure to keep track of them. Beyond that though it's just a simple recursion to see how many of each ingredient we need to spend to create an item (or rather a multiple of an item), as well as keeping track of what was spent versus is in the surplus.

Would have been 16th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count.

Time: 0:14:14.253303