2019 Day 15 Part 1
Oo, taking the robot and exploring to figure out the maze. This is a fun one.

I don't remember my thought process in 2019, but this time I pretty quickly landed on basic recursion where each level would explore and, if necessary, backtrack. This way each recursive call guarantees that the bot is back in the same spot where it originally was so it's easy to explore all 4 cardinal directions.

Then once the full maze is explored we just need to know how many steps it takes to move the bot there. lib.graph.dijkstra_length to the rescue, after a simple grid->graph conversion!

The only other thing of note is that since the Intcode program runs forever (but I don't want it to run forever) I added a new option to the VM to close the program if the input channel is closed. Off by default of course, but quite useful in this problem.

Would have been 9th, but I did the problem in 2019 so this doesn't exactly count.

Time: 0:12:12.253514