2019 Day 16 Part 1
Oh man, I'd forgotten about the Flawed Frequency Transmission problem. This one was a doozy, but more about that in part 2.

Part 1 isn't too bad, it's just a matter of understanding the transform and applying it 100 times. This could be a bit faster (and I'll probably optimize after part 2), but it gets the job done well enough.

Would have been 37th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count.

Time: 0:08:58.116983