2019 Day 16 Part 2
And part 2! Yeah, this is a doozy if you don't know what the trick is. Because the mask pattern always starts with zero we can discard any element before the answer offset. Furthermore, the offset for this problem is *deep* in the list of numbers. So deep in fact that the phase transform ends up being a reverse partial sum for the positions we care about! Partial sums are quick (unlike this FFT mask nonsense) so it's pretty easy to do 100 partial sum passes (all mod 10) and get the answer.

It took me nearly an hour to figure out the trick in 2019. In fact, I'd managed to place 113th in 2019, so honestly I did *well* back then!

Would have been 2nd on the leaderboard, but having done this problem in 2019 I already knew the trick. This *definitely* doesn't count!

Time: 0:20:05.074362