2019 Optimize Day 16
Part 1 runs way faster if Python does the summing of the ranges. (I still wouldn't want to run this too many times or on too large an array!)

Part 2 runs a little faster when I use itertools.accumulate for the partial sums. I could try to do the binomial coefficient trick, but I don't think it's really worth it. This is fast enough.