2019 Day 17 Part 1
Oh hey, it's the scaffold bot problem! Not too bad a problem, especially with the utility libraries I have. Turn the output into a string, parse it into a grid, then look for intersections.

I did lose a bit of time forgetting the API to my grid API, but not more than a minute.

Would have been 66th on the leaderboard. I lost to myself from 2019! 15 seconds slower! That's embarrassing!

Time: 0:06:46.586369