2019 Day 17 Part 2
Now for the really fun part of this problem. The first part here is extracting the sequence of steps the robot needs to take to traverse the entire scaffold. Technically there is some ambiguity in the problem (the robot could turn at any given intersection) but we just need to follow the path and turn whenever a turn is forced.

The second part is the really crazy bit though. That path needs to be compressed into a <= 20 character main program and 3 <= 20 character functions. Back in 2019 I did this step by hand initially. That's 100% the right way to do it to try to leaderboard, but this time I decided to challenge myself and solve it programatically with no hand solving.

The basic idea is a backtracking search. At any given point see if an already deduced subprogram matches the next sequence of steps that remain. If so, add that to the main program and search further. If we can't find one that matches (or if the main program gets too long and we backtrack), try and find a new subprogram (assuming we have room, we max at 3!) and see if that works. And if that fails, trim the subprogram until it's empty or we find a solution.

I did of course hit some bugs in this process as it's somewhat involved.
1) Most annoyingly, I had the wrong overall sequence to start out due to having the bot face down initially. (I always invert my y!)
2) I accidentally computed the subprogram length to be len(remaining_steps) in find_solution at one point. Unluckily it didn't hit an exception but returned a garbage solution!
3) I also got tripped up by forgetting to convert my characters back to ints to pass to the subprogram, and with forgetting to put a newline after disabling the continuous camera feed!

Still, I did programmatically solve this in a respectable amount of time. This part really screams to be solved by hand if you want to be fast.

Would have missed the leaderboard by ~5.5 minutes. In fact, I was ~6.75 minutes slower than in 2019, but I did this part by hand in 2019. I'm not too mad about my showing here :)

Time: 0:50:47.789638