2019 Day 18 Part 1
This problem really hurt me in 2019. The only problem I ever slept on (excluding 2015 Day 19 which I didn't do live). In fact, much of my lib.graph library was born *out* of this problem!

The simple stuff is the grid parsing. I even have a library to help with that!

The hard part is the actual searching through the grid. The state space is designed to be exponential for a naive search but there are some key observations that allow for drastic search improvements:
1) No matter what keys you have the optimal path between two points of interest doesn't change. (You may have to go *through* a point of interest to reach another one, but that's okay.) This lets one construct a minimized point of interest connectivity graph.
2) Two states are identical if the bot is in the same location and has the same keys. It doesn't matter what order the keys were grabbed in! As such, always keep a sorted list of keys!

With these in mind a neighbor function can be written for any state. It searches through the connectivity graph (faster than using the grid directly) and returns any valid neighbor, adjusting the key list as necessary. Any locked doors are excluded since they aren't traversable so they can't be part of the best path.

With all that out of the way it's just a matter of finding the first state with all they keys. (We don't care where the bot is, so it's a fuzzy match!)

Would have been 22nd on the leaderboard but since I did this in 2019 (and really internalized it from how much I struggled with it) this definitely doesn't count. In fact, I beat my 2019 time by just over 8 hours!

Time: 0:18:41.062761