2019 Day 18 Part 2
Part 2 really isn't much of a jump over part 1, to be honest. Keeping track of 4 bots instead of 1 just complicates the state record and neighbor function a bit.

Admittedly this implementation is a bit slow (bots that can't make progress keep roaming about in the search!) but it was faster to let this run to completion than to optimize it. I'll optimize it in a follow-up, shouldn't be too hard at all.

Would have been a crazy 3rd on the leaderboard but since I did this in 2019 (and really internalized it from how much I struggled with it) this definitely doesn't count. In fact, I beat my 2019 time by just over 8 hours!

Time: 0:31:21.085721