2019 Day 19 Part 1
Oh hey, it's the tractor beam problem. Sum the affected coordinates in the first 50x50 square, that's not too bad.

I did get tripped up by not realizing that we had to restart the Intcode program for each query. Cloning the Intcode VM to try to save time...didn't really save time. It still takes nearly 2 seconds to run. However, it does run fine.

Would have been 34th, 35th, 36th, or 37th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count. That being said, I was 43rd on the leaderboard in 2019 (only 8-9 seconds slower) so that's pretty close!

Time: 0:02:35.750974