2019 Day 19 Part 2
Now we need to find the first 100x100 square fully affected by the tractor beam. I think that this is the same approach I used in 2019: Check the extreme corners and nudge the search square until the extreme corners are hit. Then do a final check on the full 100x100 square to ensure that it is indeed affected.

This does take a while to run, a full 55 seconds! I can probably optimize the search, and maybe optimize my Intcode VM. I don't like how long this takes to run at all!

Would have been 7th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count. I did beat my 2019 time by 22-23 minutes though!

Time: 0:10:50.253317