2019 Day 20 Part 1
Oh, the recursive maze problem was 2019? Fun! Well, the first problem with this problem is figuring out where the portals are. This was *so much* easier than in 2019, it's not even funny. I did have some trouble due to the labels always being top-down or left-right, but that was easy to correct for.

Once I got the labels figured out it was pretty easy to convert to a graph. From there lib.graph.dijkstra_length gets the answer trivially!

Would have been 83rd, 84th, or 85th, but I did the problem in 2019 so this doesn't exactly count. I did beat my 2019 time by ~11 minutes though! Having an easier time parsing really helped!

Time: 0:23:58.231146