2019 Day 20 Part 2
Oh right, I need to know the difference between inner and outer portals. I forgot about that! Well, at least that wasn't hard to figure out. (Just figure out where the outer portals can be and it's easy to disambiguate.)

Anyway, once the base graph is determined (with knowledge of which side each portal is on) it's a simple deal of dikjstra with a name, level state. The entire complication here was just the logic in the recursive_neighbor function!

Would have been 36th on the leaderboard, but I did the problem in 2019 so this doesn't exactly count. Still, that's a good ranking! In fact, I finished nearly an hour faster than I did in 2019. I forget why that took so long...

Time: 0:37:29.702760