2019 Day 21 Part 2
And now we run. More sensors means we can look ahead and make sure that we can land if we need to jump twice. Technically I didn't do this 100% correctly (I should have checked if we can land, step once, then jump again) but it works.

Would have been 14th on the leaderboard but I did the problem in 2019 so this doesn't exactly count.

Time: 0:17:10.967735