2019 Day 22 Part 2
And here's part 2! I got to skip a bunch of the discoveries this time around and had a much nicer time solving this.

First off, the shuffle operations can all be represented by modular formulae. These in turn can be composed to get a single RNG formula to perform a single shuffle. Then once that's done the RNG can be "raised" to a power (applied n times) to make a super RNG to apply the shuffle N times. This can be done with fast exponentiation because I certainly don't have time to wait for ~102 trillion applications!

Finally, the trickiest part. We need to go back in time! We want to know what card ends up in position 2020. There is a clever approach that only uses forward applications of the RNG (I think that it cycles after DECK_SIZE-1 applications) but I don't really like that approach. Instead, the RNG formula needs to be inverted. Back in 2019 this was a pain since I didn't know how to get a modular multiplicative inverse (still don't really, I'm not super comfortable with the extended euclidian algorithm). But not to worry, Python knows how to do that for me! Once I get that, inverting the RNG is easy. Then the reverse RNG directly gives the right answer! Yay!

Would have been 40th on the leaderboard (or something like 22nd if I hadn't wasted 11 minutes in part 1) but I did the problem in 2019 and *really* learned from it so this definitely doesn't count. Still though, I'll take that showing!

Time: 0:57:39.341757