2019 Day 23 Part 1
Here we go, the networking problem. This is *the* reason I wanted a lib.channels library. This gave me so much trouble in go in 2019!

There's a good bit of marshaling and management going on here, mostly to help keep things straight. But overall this is still a relatively straightforward implementation of the problem specification.

The big trick I employed here was *always* sending a -1 after every packet sent to the programs. The problem said only to do that when nothing was in the queue, but honestly it's kind of hard to do that in my architecture. The programs can just be told that nothing is available a bunch and sort themselves out.

Would have missed the leaderboard by ~6 minutes, but I'm fine with that. This networking emulation stuff is actually kind of tricky to get right without deadlock. (Especially since I'm running the Intcode VM on a background thread!)

Time: 0:17:56.286756