2019 Day 23 Part 2
Oof, this was the reason I had deadlock detection in my channel library yet I still had trouble! Long story short, I included too few channels in the deadlock detection and was informed of deadlock sporadically even when it didn't happen.

That aside, this also went kind of smoothly. There's a ton of duplication with part 1 and the code is cluttered, but it works well and is stable! (How many people actually dealt with real threading issues in this problem, I wonder?)

Would have missed the leaderboard by ~19 minutes! But still, this was way smoother than 2019 where it took 85 minutes and was still ridden with race conditions!

Time: 0:39:37.915586