2019 Refactor Day 23
There isn't a lot of difference between parts 1 and 2. It's mostly in the controller loop which can be parameterized with an include_nat flag. Doing so allows me to extract a common run() function used by both parts 1 and 2.

Furthermore, I can get rid of the hacky -1 signals I send all the time by keeping track of which machines are active. Whenever we get a deadlock, send a -1 to one of the machine that was last seen active. If that wakes up the network then great! If we run out of machines to poke then we're in absolute deadlock and the NAT needs to jump into action.

Overall this feels like a better approach than the -1 hack I did before, probably more true to real networking. Plus it facilitated the deduplication.

This is so much nicer than my 2019 version!