2019 Day 24 Part 2
And now the bugs went recursive on us! This is still essentially a game of life variant, but now it's unbounded with a far more interesting ruleset.

Speaking of sets, using sets to solve a problem like this is very smart. That way it's easy to iterate over bugs and count which neighbor cells they touch without needing to worry about what the boundaries are, they'll work themselves out! In fact, this was super easy to code up...except for me completely flubbing my inner cell check! For some reason I was checking if (nx, ny) == 2 instead of (nx, ny) == (2, 2). This meant that no bugs thought that they recursed down! Once that was fixed everything just worked, but ugh.

Would have been 21st on the leaderboard. Would have been 7th had I not had that dumb bug wasting 7 and a half minutes! (That being said, I did the problem in 2019 so this doesn't exactly count.)

Time: 0:27:01.381947