2019 Day 25 Part 1 (There is no Part 2)
It's the text adventure! It's nice not needing to figure out how to actually get console IO working like I had to with go. Not only that but I was able to pretty easily set up an initial commands structure in Python and use that whenever I needed to restart the adventure. (Which was a lot.)

After that it was just a matter of testing all item combinations. Unlike 2019 I did this programmatically. Just take the 8 items and go through the 256 combinations of holding or not holding them. Comparing bit masks helps me minimize item transitions between attempts at the door which is good. I don't even bother trying to figure out which one was right! I just queue up all the commands and let the program terminate whenever one of them works!

I'll definitely want to come back to this and *truly* solve it programmatically. But this was nice! I finished in about half the time it took me in 2019!

Would have been 40th or 41st on the leaderboard but I did the problem in 2019 so this doesn't exactly count.
Then would have been 35th to get the second star. Again, doesn't really count.

Time: 0:21:41.274556